third commit

+++ contracts/Crowdfunding.sol
@@ -11,7 +11,7 @@ contract Crowdfunding {
     constructor(
         string memory _name,
         string memory _description,
-        uint _goal,
+        uint256 _goal,
         uint256 _durationInDays
     ){
         name = _name;
@@ -19,5 +19,6 @@ contract Crowdfunding {
         goal = _goal;
         deadline = block.timestamp + (_durationInDays * 1 days);
         owner = msg.sender;
+        
     }
 }
